chore: get rid of custom plugins in visual test config
