+++ viewer/package.json
@@ -50,7 +50,7 @@
     "test": "yarn run test:wasm && yarn run test:ts",
     "test:ts": "yarn run build:wasm-test && vitest run --reporter=verbose",
     "test:wasm": "yarn workspaces foreach -A run run-wasm-pack test --headless --chrome ./wasm",
-    "test:visual:server": "yarn vite dev --config ./packages/visual-tests.vite.config.ts",
+    "test:visual:server": "yarn vite dev --config ./visual-tests/visual-tests.vite.config.ts",
     "test:visual": "yarn run build:wasm && playwright test --config=visual-tests/playwright.config.ts",
     "test:visual:update": "playwright test --config=visual-tests/playwright.config.ts --update-snapshots",
     "typecheck:lib": "tsc --project tsconfig.lib.json --noEmit",
@@ -58,7 +58,7 @@
     "coverage": "yarn run build:wasm-test && vitest run --coverage",
     "lint": "eslint . --ext .ts,.js --max-warnings 0 --cache",
     "update-api": "api-extractor run --local && shx rm -rf api-extractor-temp/",
-    "ws:serve": "yarn run build:wasm && yarn vite dev --config ./packages/visual-tests.vite.config.ts",
+    "ws:serve": "yarn run build:wasm && yarn vite dev --config ./visual-tests/visual-tests.vite.config.ts",
     "ws:test": "cd $INIT_CWD && vitest run",
     "ws:test:wasm": "cd $INIT_CWD && wasm-pack test --chrome --headless ./wasm",
     "ws:update-cargo-index": "cargo update --dry-run",
