+++ viewer/tsconfig.lib.json
@@ -5,7 +5,7 @@
     "stripInternal": true,
     "types": ["vite/client", "node", "vite-plugin-glsl/ext"]
   },
-  "include": ["./packages/**/*.ts", "./api-entry-points/**/*.ts"],
+  "include": ["./packages/**/*.ts", "./api-entry-points/**/*.ts", "./visual-tests/visual-tests.vite.config.ts"],
   "exclude": [
     "./packages/**/*.d.ts",
     "./packages/**/*.spec.ts",
