+++ viewer/visual-tests/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8" />
+</head>
+<body>
+<script type="module" src="/VisualTest.browser.ts"></script>
+</body>
+</html>
