+++ viewer/visual-tests/public/pointcloud-bunny/ept-hierarchy/0-0-0-0.json
@@ -1,0 +1,3 @@
+{
+  "0-0-0-0": 565
+}
