+++ viewer/visual-tests/public/primitives/360Images.json
@@ -1,0 +1,154 @@
+[
+  {
+    "translation": {
+      "x": 5,
+      "y": 3,
+      "z": -10
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "faces": [
+      {
+        "face": "back",
+        "id": 0
+      },
+      {
+        "face": "bottom",
+        "id": 1
+      },
+      {
+        "face": "front",
+        "id": 2
+      },
+      {
+        "face": "left",
+        "id": 3
+      },
+      {
+        "face": "right",
+        "id": 4
+      },
+      {
+        "face": "top",
+        "id": 5
+      }
+    ]
+  },
+  {
+    "translation": {
+      "x": 10,
+      "y": 3,
+      "z": -5
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "faces": [
+      {
+        "face": "back",
+        "id": 0
+      },
+      {
+        "face": "bottom",
+        "id": 1
+      },
+      {
+        "face": "front",
+        "id": 2
+      },
+      {
+        "face": "left",
+        "id": 3
+      },
+      {
+        "face": "right",
+        "id": 4
+      },
+      {
+        "face": "top",
+        "id": 5
+      }
+    ]
+  },
+  {
+    "translation": {
+      "x": 10,
+      "y": 3,
+      "z": -10
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "faces": [
+      {
+        "face": "back",
+        "id": 0
+      },
+      {
+        "face": "bottom",
+        "id": 1
+      },
+      {
+        "face": "front",
+        "id": 2
+      },
+      {
+        "face": "left",
+        "id": 3
+      },
+      {
+        "face": "right",
+        "id": 4
+      },
+      {
+        "face": "top",
+        "id": 5
+      }
+    ]
+  },
+  {
+    "translation": {
+      "x": 5,
+      "y": 3,
+      "z": -5
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "faces": [
+      {
+        "face": "back",
+        "id": 0
+      },
+      {
+        "face": "bottom",
+        "id": 1
+      },
+      {
+        "face": "front",
+        "id": 2
+      },
+      {
+        "face": "left",
+        "id": 3
+      },
+      {
+        "face": "right",
+        "id": 4
+      },
+      {
+        "face": "top",
+        "id": 5
+      }
+    ]
+  }
+]
