+++ viewer/visual-tests/public/primitives/scene.json
@@ -1,0 +1,1 @@
+{"version":9,"projectId":123,"modelId":456,"revisionId":789,"subRevisionId":999,"maxTreeIndex":76,"sectors":[{"id":0,"sectorFileName":"0.glb","path":"0/","depth":0,"boundingBox":{"min":{"x":-1.3934096876224187,"y":-1,"z":-1.788780573085674},"max":{"x":28.29323544313644,"y":25.458996612973937,"z":1.6256651949082248}},"geometryBoundingBox":{"min":{"x":-1.3934096876224187,"y":-1,"z":-1.788780573085674},"max":{"x":28.29323544313644,"y":25.458996612973937,"z":1.6256651949082248}},"estimatedTriangleCount":11632,"estimatedDrawCallCount":17,"minDiagonalLength":2,"maxDiagonalLength":2.8284270763397217,"downloadSize":27928}]}
