+++ viewer/visual-tests/public/textured/scene.json
@@ -1,0 +1,1 @@
+{"version":9,"projectId":123,"modelId":456,"revisionId":789,"subRevisionId":999,"maxTreeIndex":11,"unit":"Meters","sectors":[{"id":0,"sectorFileName":"0.glb","texturedFileName":"0_textured.glb","path":"0/","depth":0,"boundingBox":{"min":{"x":-111.28918838500977,"y":-139.4479528269494,"z":-505.1957135996545},"max":{"x":461.8326663970947,"y":183.60158975932245,"z":482.5748945078576}},"geometryBoundingBox":{"min":{"x":-111.28918838500977,"y":-139.4479528269494,"z":-505.1957135996545},"max":{"x":461.8326663970947,"y":183.60158975932245,"z":482.5748945078576}},"estimatedTriangleCount":38,"estimatedDrawCallCount":1,"minDiagonalLength":346.4101867675781,"maxDiagonalLength":346.4101867675781,"downloadSize":2876}]}
